✨ feat: Add new import statement for "Flexbox" component in "Empty" component
The code changes add a new import statement for the "Flexbox" component from the "react-layout-kit" library in the "Empty" component. This allows the usage of the "Flexbox" component in the "Empty" component.

feat: Add new "Flexbox" component wrapping existing content in "Empty" component

The code changes include adding a new "Flexbox" component in the "Empty" component, which wraps the existing content. This provides a flexible layout for the content in the "Empty" component.

style: Add new CSS class "image" with "align-self: center" property in "Empty" style file

The code changes in the "Empty" style file add a new CSS class called "image" with the "align-self: center" property. This aligns the image content in the "Empty" component to the center.

style: Change backdrop-filter blur value from 5px to 2px in "PluginDevModal" component

The code changes in the "PluginDevModal" component modify the backdrop-filter blur value from 5px to 2px. This adjusts the visual appearance of the modal backdrop.

feat: Add "enableShortKey" and "shortKey" props to "SearchBar" component in "SessionList/Header" component

The code changes in the "SessionList/Header" component include adding the "enableShortKey" prop and a new "shortKey" prop with the value "k" to the "SearchBar" component. This enables the usage of keyboard shortcuts for the search functionality in the header component.

Summary of changes:
- Added import statement for "Flexbox" component in "Empty" component
- Added new "Flexbox" component wrapping existing content in "Empty" component
- Added CSS class "image" with "align-self: center" property in "Empty" style file
- Changed backdrop-filter blur value from 5px to 2px in "PluginDevModal" component
- Added "enableShortKey" and "shortKey" props to "SearchBar" component in "SessionList/Header" component